[ch56784] check if changes are compatible with python2
