+++ custom-recipes/timeseries-preparation-resampling/recipe.json
@@ -56,7 +56,7 @@
       "type": "COLUMN",
       "columnRole": "input_dataset",
       "mandatory": false,
-      "description": "⚠️  Deprecated",
+      "description": "⚠️  Deprecated. Use the field 'Column with identifiers' instead",
       "visibilityCondition": "model.advanced_activated && model.groupby_column"
     },
     {
